gateway/requirements.txt
aiohttp==3.8.1
grpcio-tools==1.47.0
PyJWT==2.3.0
